interfaces reorganization
